+++ src/views/viewer/ViewerLayout.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
 import { ErrorBoundary } from '@sentry/react'
-import React, { useEffect, useRef, useState } from 'react'
-import { Route, Routes, useLocation, useNavigate, useNavigationType } from 'react-router-dom'
+import React from 'react'
+import { Route, Routes, useLocation, useNavigate } from 'react-router-dom'
 import { CSSTransition, SwitchTransition } from 'react-transition-group'
 
 import { ViewErrorBoundary } from '@/components/ViewErrorFallback'
@@ -26,8 +26,6 @@ import {
 
 import { DiscoverView } from './DiscoverView/DiscoverView'
 
-const ROUTING_ANIMATION_OFFSET = 100
-
 const viewerRoutes = [
   { path: relativeRoutes.viewer.search(), element: <SearchView /> },
   { path: relativeRoutes.viewer.index(), element: <HomeView /> },
@@ -43,33 +41,10 @@ const viewerRoutes = [
 export const ViewerLayout: React.FC = () => {
   const location = useLocation()
   const locationState = location.state as RoutingState
-  const scrollPosition = useRef<number>(0)
 
   const navigate = useNavigate()
-  const navigationType = useNavigationType()
-  const [cachedLocation, setCachedLocation] = useState(location)
   const mdMatch = useMediaMatch('md')
   const { searchOpen } = useSearchStore()
-
-  useEffect(() => {
-    if (location.pathname === cachedLocation.pathname) {
-      return
-    }
-
-    setCachedLocation(location)
-
-    if (locationState?.overlaidLocation?.pathname === location.pathname) {
-      // if exiting routing overlay, skip scroll to top
-      return
-    }
-    if (navigationType !== 'POP') {
-      scrollPosition.current = window.scrollY
-    }
-    // delay scroll to allow transition to finish first
-    setTimeout(() => {
-      window.scrollTo(0, navigationType !== 'POP' ? 0 : scrollPosition.current)
-    }, parseInt(transitions.timings.routing) + ROUTING_ANIMATION_OFFSET)
-  }, [location, cachedLocation, locationState, navigationType])
 
   const displayedLocation = locationState?.overlaidLocation || location
